requirements.txt
# Core dependencies
PyQt5==5.15.9
numpy==1.24.3
praw==7.7.1
google-generativeai==0.3.1
whisper==1.1.10
soundfile==0.12.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
diffusers==0.21.4
transformers==4.31.0
accelerate==0.21.0
moviepy==1.0.3
Pillow==10.0.0
tqdm==4.66.1

# Optional dependencies (uncomment if needed)
# kokoro==0.1.3  # For TTS, may need to be installed separately
# safetensors==0.3.2  # For optimized model loading 